forecast.py
This function takes two arguments: a list of demand data and the length of the time period covered by the data (in weeks). It calculates the average demand for the given period and then adjusts the forecast based on the trend in the data. If the demand has been increasing over the given period, the forecast is increased by 10%. If the demand has been decreasing, the forecast is decreased by 10%. If the demand has remained stable, the forecast is equal to the average demand.